fix: use bubbleWhere on leaderboards
