+++ src/services/has-many-getter.js
@@ -1,6 +1,5 @@
-
+import QueryUtils from '../utils/query';
 import Operators from '../utils/operators';
-import Recursion from '../utils/recursion';
 import CompositeKeysManager from './composite-keys-manager';
 import ResourcesGetter from './resources-getter';
 
@@ -50,38 +49,10 @@ class HasManyGetter extends ResourcesGetter {
       parentOptions.order = (options.order || []).map((o) => [associationName, ...o]);
     }
 
-    this._bubbleWheresInPlace(parentOptions, associationName);
+    const operators = Operators.getInstance({ Sequelize: this._parentModel.sequelize.constructor });
+    QueryUtils.bubbleWheresInPlace(parentOptions, operators);
 
     return parentOptions;
-  }
-
-  /**
-   * Extract all where conditions along the include tree, and bubbles them up to the top.
-   * This allows to work around a sequelize quirk that cause nested 'where' to fail when they
-   * refer to relation fields from an intermediary include (ie '$book.id$').
-   *
-   * This happens when forest admin filters on relations are used.
-   *
-   * @see https://sequelize.org/master/manual/eager-loading.html#complex-where-clauses-at-the-top-level
-   * @see https://github.com/ForestAdmin/forest-express-sequelize/blob/7d7ad0/src/services/filters-parser.js#L104
-   */
-  _bubbleWheresInPlace(options) {
-    const Ops = Operators.getInstance({ Sequelize: this._parentModel.sequelize.constructor });
-
-    (options.include ?? []).forEach((include) => {
-      this._bubbleWheresInPlace(include);
-
-      const newWhere = Recursion.mapKeysDeep(include.where, (key) => (
-        key[0] === '$' && key[key.length - 1] === '$'
-          ? `$${include.as}.${key.substring(1)}`
-          : `$${include.as}.${key}$`
-      ));
-
-      delete include.where;
-      if (newWhere) {
-        options.where = { [Ops.AND]: [options.where, newWhere] };
-      }
-    });
   }
 }
